Move some commitment related functions
