+++ packages/wallet/src/redux/indirect-funding/player-a/reducer.ts
@@ -4,12 +4,10 @@ import * as actions from '../../actions';
 import { unreachable } from '../../../utils/reducer-utils';
 import { PlayerIndex, WalletProcedure } from '../../types';
 
-import { Channel, CommitmentType } from 'fmg-core';
+import { Channel } from 'fmg-core';
 import { channelID } from 'magmo-wallet-client/node_modules/fmg-core/lib/channel';
 import * as channelState from '../../channel-state/state';
 import { Commitment } from 'fmg-core/lib/commitment';
-import { signCommitment } from '../../../utils/signing-utils';
-import { bytesFromAppAttributes, appAttributesFromBytes } from 'fmg-nitro-adjudicator';
 import { channelStateReducer } from '../../channel-state/reducer';
 import * as channelActions from '../../channel-state/actions';
 import { messageRelayRequested } from 'magmo-wallet-client';
@@ -19,6 +17,12 @@ import { CHANNEL_FUNDED } from '../../funding-state/state';
 import * as channelStates from '../../channel-state/state';
 import { addHex } from '../../../utils/hex-utils';
 import { accumulateSideEffects } from '../../outbox';
+import {
+  hasConsensusBeenReached,
+  composeLedgerUpdateCommitment,
+  composePostFundCommitment,
+  composePreFundCommitment,
+} from '../../../utils/commitment-utils';
 
 export function playerAReducer(
   state: walletStates.Initialized,
@@ -320,103 +324,4 @@ export const updateFundingState = (
 ) => {
   const updatedFundingState = fundingStateReducer(state.fundingState, action);
   state.fundingState = updatedFundingState.state;
-};
-
-export const hasConsensusBeenReached = (
-  lastCommitment: Commitment,
-  penultimateCommitment: Commitment,
-) => {
-  const numOfPlayers = lastCommitment.channel.participants.length;
-  const lastAppAttributes = appAttributesFromBytes(lastCommitment.appAttributes);
-  const penultimateAppAttributes = appAttributesFromBytes(penultimateCommitment.appAttributes);
-
-  if (
-    lastAppAttributes.consensusCounter === numOfPlayers - 1 &&
-    lastCommitment.allocation === penultimateAppAttributes.proposedAllocation &&
-    lastCommitment.destination === penultimateAppAttributes.proposedDestination
-  ) {
-    return true;
-  } else {
-    return false;
-  }
-};
-
-// Commitment composers
-
-export const composeLedgerUpdateCommitment = (
-  lastCommitment: Commitment,
-  ourIndex: PlayerIndex,
-  proposedAllocation: string[],
-  proposedDestination: string[],
-  privateKey: string,
-) => {
-  const { channel, turnNum: previousTurnNum, allocation, destination } = lastCommitment;
-  const appAttributes = bytesFromAppAttributes({
-    proposedAllocation,
-    proposedDestination,
-    consensusCounter: ourIndex,
-  });
-  const updateCommitment: Commitment = {
-    channel,
-    commitmentType: CommitmentType.PostFundSetup,
-    turnNum: previousTurnNum + 1,
-    commitmentCount: ourIndex,
-    allocation,
-    destination,
-    appAttributes,
-  };
-  const commitmentSignature = signCommitment(updateCommitment, privateKey);
-
-  return { updateCommitment, commitmentSignature };
-};
-
-export const composePostFundCommitment = (
-  lastCommitment: Commitment,
-  ourIndex: PlayerIndex,
-  privateKey: string,
-) => {
-  const {
-    channel,
-    turnNum: previousTurnNum,
-    allocation,
-    destination,
-    appAttributes,
-  } = lastCommitment;
-  const postFundCommitment: Commitment = {
-    channel,
-    commitmentType: CommitmentType.PostFundSetup,
-    turnNum: previousTurnNum + 1,
-    commitmentCount: ourIndex,
-    allocation,
-    destination,
-    appAttributes,
-  };
-  const commitmentSignature = signCommitment(postFundCommitment, privateKey);
-
-  return { postFundCommitment, commitmentSignature };
-};
-export const composePreFundCommitment = (
-  channel: Channel,
-  allocation: string[],
-  destination: string[],
-  ourIndex: PlayerIndex,
-  privateKey: string,
-) => {
-  const appAttributes = bytesFromAppAttributes({
-    proposedAllocation: allocation,
-    proposedDestination: destination,
-    consensusCounter: ourIndex,
-  });
-  const preFundSetupCommitment: Commitment = {
-    channel,
-    commitmentType: CommitmentType.PreFundSetup,
-    turnNum: ourIndex,
-    commitmentCount: ourIndex,
-    allocation,
-    destination,
-    appAttributes,
-  };
-  const commitmentSignature = signCommitment(preFundSetupCommitment, privateKey);
-
-  return { preFundSetupCommitment, commitmentSignature };
 };
